+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixes:
 
+* [#24](https://github.com/MiSchroe/klf-200-api/issues/24) Add `refreshScenes()` method to re-read scenes data from the gateway
 * [#27](https://github.com/MiSchroe/klf-200-api/issues/27) Add `refresh()` method to re-read object data from the gateway
 * Fix security audit dependencies ([#32](https://github.com/MiSchroe/klf-200-api/issues/32), [#35](https://github.com/MiSchroe/klf-200-api/issues/35), [#37](https://github.com/MiSchroe/klf-200-api/issues/37))